+++ com.unity.render-pipelines.high-definition/Editor/Material/Unlit/HDShaderGUI.cs
@@ -25,31 +25,13 @@ namespace UnityEditor.Rendering.HighDefinition
 
         public override void AssignNewShaderToMaterial(Material material, Shader oldShader, Shader newShader)
         {
+            // When switching shader, the custom RenderQueue is reset due to shader assignment
+            // To keep the correct render queue we need to save it here, do the change and re-assign it
+            int currentRenderQueue = material.renderQueue;
             base.AssignNewShaderToMaterial(material, oldShader, newShader);
-
-            ResetMaterialCustomRenderQueue(material);
+            material.renderQueue = currentRenderQueue;
 
             SetupMaterialKeywordsAndPassInternal(material);
-        }
-
-        protected static void ResetMaterialCustomRenderQueue(Material material)
-        {
-            HDRenderQueue.RenderQueueType targetQueueType;
-            switch (material.GetSurfaceType())
-            {
-                case SurfaceType.Opaque:
-                    targetQueueType = HDRenderQueue.GetOpaqueEquivalent(HDRenderQueue.GetTypeByRenderQueueValue(material.renderQueue));
-                    break;
-                case SurfaceType.Transparent:
-                    targetQueueType = HDRenderQueue.GetTransparentEquivalent(HDRenderQueue.GetTypeByRenderQueueValue(material.renderQueue));
-                    break;
-                default:
-                    throw new ArgumentException("Unknown SurfaceType");
-            }
-
-            float sortingPriority = material.GetFloat(kTransparentSortPriority);
-            bool alphaTest = material.GetFloat(kAlphaCutoffEnabled) > 0.5f;
-            material.renderQueue = HDRenderQueue.ChangeType(targetQueueType, (int)sortingPriority, alphaTest);
         }
         
         readonly static string[] floatPropertiesToSynchronize = {