+++ com.unity.render-pipelines.high-definition/Runtime/Lighting/Light/HDAdditionalLightData.Migration.cs
@@ -89,6 +89,7 @@ namespace UnityEngine.Rendering.HighDefinition
                         data.shadowFadeDistance = additionalShadow.shadowFadeDistance;
                         data.shadowTint = additionalShadow.shadowTint;
                         data.normalBias = additionalShadow.normalBias;
+                        data.constantBias = additionalShadow.constantBias;
                         data.shadowUpdateMode = additionalShadow.shadowUpdateMode;
                         data.shadowCascadeRatios = additionalShadow.shadowCascadeRatios;
                         data.shadowCascadeBorders = additionalShadow.shadowCascadeBorders;
@@ -153,12 +154,12 @@ namespace UnityEngine.Rendering.HighDefinition
             SetEmissiveMeshRendererEnabled(true);
         }
 
+        //Migration function called in:
+        // - OnValidate()
+        // - Awake()
+        // both function are called at instance
         void Migrate()
-        {
-            k_HDLightMigrationSteps.Migrate(this);
-            // OnValidate might be called before migration but migration is needed to call UpdateBounds() properly so we call it again here to make sure that they are updated properly.
-            OnValidate();
-        }
+            => k_HDLightMigrationSteps.Migrate(this);
 
         void Awake() => Migrate();