+++ com.unity.render-pipelines.high-definition/Runtime/Sky/SkySettings.cs
@@ -38,6 +38,21 @@ namespace UnityEngine.Rendering.HighDefinition
         ShowUpdateMode =      (1 << 2),
     }
 
+    public enum BackplateType
+    {
+        Disc,
+        Rectangle,
+        Ellipse,
+        Infinite
+    }
+
+    [Serializable, DebuggerDisplay(k_DebuggerDisplay)]
+    public sealed class BackplateTypeParameter : VolumeParameter<BackplateType>
+    {
+        public BackplateTypeParameter(BackplateType value, bool overrideState = false)
+            : base(value, overrideState) { }
+    }
+
     [Serializable, DebuggerDisplay(k_DebuggerDisplay)]
     public sealed class SkyIntensityParameter : VolumeParameter<SkyIntensityMode>
     {
@@ -57,6 +72,8 @@ namespace UnityEngine.Rendering.HighDefinition
         public MinFloatParameter        multiplier = new MinFloatParameter(1.0f, 0.0f);
         [Tooltip("Informative helper that displays the relative intensity (in Lux) for the current HDR texture set in HDRI Sky.")]
         public MinFloatParameter        upperHemisphereLuxValue = new MinFloatParameter(1.0f, 0.0f);
+        [Tooltip("Informative helper that displays Show the color of Shadow.")]
+        public Vector3Parameter         upperHemisphereLuxColor = new Vector3Parameter(new Vector3(0, 0, 0));
         [Tooltip("Sets the absolute intensity (in Lux) of the current HDR texture set in HDRI Sky. Functions as a Lux intensity multiplier for the sky.")]
         public FloatParameter           desiredLuxValue = new FloatParameter(20000);
         [Tooltip("Specifies when HDRP updates the environment lighting. When set to OnDemand, use HDRenderPipeline.RequestSkyEnvironmentUpdate() to request an update.")]